Refactors maxProductOfThree(). Fixes #41 (#43)
Refactors by sorting the array (ascending) and then taking the maximal product from the three last numbers and the last number and the two first numbers, instead of looping through all numbers and "manually" setting the three largest and two smallest numbers (which requires a lot of if-statements). The CC is reduced from 11 to 2.